@@ -1,39 +1,39 @@
 name: upgrader
 description: Flutter package for prompting users to upgrade when there is a newer version of the app in the store.
-version: 7.0.0
+version: 9.0.0
 homepage: https://github.com/larryaasen/upgrader
 
 environment:
-  sdk: '>=2.18.6 <4.0.0'
-  flutter: ">=3.10.0"
+  sdk: '>=3.1.0 <4.0.0'
+  flutter: ">=3.13.1"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # From fluttercommunity.dev: Get current device information from within the Flutter application.
-  device_info_plus: ^9.0.0
+  device_info_plus: '>=8.1.0 <10.0.0'
 
   # Pure Dart library for HTML5 parsing
   html: ^0.15.3
 
   # For app_theme.dart
-  google_fonts: ^2.3.1
+  google_fonts: ^6.2.1
 
   # Determine if it is HMS device
   google_api_availability: ^3.0.1
 
   # From Dart Team: A composable, Future-based library for making HTTP requests.
-  http: ^0.13.6
+  http: '>=0.13.6 <2.0.0'
 
   # From dart.dev: Platform independent access to information about the current operating system.
   os_detect: ^2.0.1
 
   # From Flutter Community: Provides an API for querying information about an application package.
-  package_info_plus: ^4.0.0
+  package_info_plus: '>=4.0.1 <6.0.0'
 
   # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
-  shared_preferences: ^2.1.1
+  shared_preferences: '>=2.1.1 <3.0.0'
 
   # From Flutter Team: A Flutter plugin for launching a URL in the mobile platform.
   url_launcher: ^6.1.11
@@ -49,7 +49,8 @@ dev_dependencies:
     sdk: flutter
 
   # From Dart Team: Mock library for Dart inspired by Mockito.
-  mockito: ^5.4.0
-  flutter_lints: ^2.0.1
+  mockito: ^5.4.3
+  flutter_lints: ^2.0.3
+  go_router: ^7.1.1
 
 flutter:
