@@ -19,9 +19,15 @@ import 'package:test/test.dart';
 import 'package:path/path.dart' as path_d;
 
 void main() {
-  group('Driver Test:', () {
-    late FlutterDriver driver;
+  late FlutterDriver driver;
 
+  Future<void> verifyText(String text) async {
+    print('verify start: $text');
+    await driver.waitFor(find.text(text));
+    print('verify end');
+  }
+
+  group('Driver Test:', () {
     // Connect to the Flutter driver before running any tests
     setUpAll(() async {
       print('pwd: ${Directory.current.path}');
@@ -43,15 +49,39 @@ void main() {
     });
 
     test('verify app started', () async {
-      // await takeScreenshot(driver, 'app_startup.png');
+      print('test started');
+      await driver.waitFor(find.text('Upgrader Driver App'));
+      await driver.waitFor(find.text('Dialog Alert'));
+      await driver.waitFor(find.text('Dialog Alert - Cupertino'));
+      print('test ended');
+    });
 
-      await driver.waitFor(find.text('Upgrader Example'));
+    test('verify alert', () async {
+      print('test started');
+      await driver.tap(find.text('Dialog Alert'));
       await driver.waitFor(find.text('Update App?'));
       await driver.waitFor(find.text('Would you like to update it now?'));
       await driver.waitFor(find.text('Release Notes'));
       await driver.waitFor(find.text('IGNORE'));
       await driver.waitFor(find.text('LATER'));
       await driver.waitFor(find.text('UPDATE NOW'));
+      await driver.tap(find.text('IGNORE'));
+      print('test ended');
+    });
+
+    test('verify alert - cupertino', () async {
+      print('test started');
+      await driver.tap(find.text('Dialog Alert - Cupertino'));
+
+      await verifyText('Update App?');
+      await verifyText('Would you like to update it now?');
+      await verifyText('Release Notes');
+      await verifyText('IGNORE');
+      await verifyText('LATER');
+      await verifyText('UPDATE NOW');
+
+      await driver.tap(find.text('IGNORE'));
+      print('test ended');
     });
   });
 }
