@@ -1,16 +1,16 @@
 /*
- * Copyright (c) 2018-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2023 Larry Aasen. All rights reserved.
  */
 
 import 'dart:async';
+import 'dart:io';
+import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:google_api_availability/google_api_availability.dart';
 import 'package:http/http.dart' as http;
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:upgrader/src/app_update_dialog.dart';
 import 'package:upgrader/src/hms_search_api.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:version/version.dart';
@@ -18,7 +18,7 @@ import 'package:version/version.dart';
 import 'appcast.dart';
 import 'itunes_search_api.dart';
 import 'play_store_search_api.dart';
-import 'upgrade_io.dart';
+import 'upgrade_os.dart';
 import 'upgrade_messages.dart';
 
 /// Signature of callbacks that have no arguments and return bool.
@@ -35,12 +35,12 @@ typedef WillDisplayUpgradeCallback = void Function(
     String? installedVersion,
     String? appStoreVersion});
 
-/// There are two different dialog styles: Cupertino and Material
-enum UpgradeDialogStyle { cupertino, material }
+/// The type of data in the stream.
+typedef UpgraderEvaluateNeed = bool;
 
 /// A class to define the configuration for the appcast. The configuration
 /// contains two parts: a URL to the appcast, and a list of supported OS
-/// names, such as "android", "ios".
+/// names, such as "android", "fuchsia", "ios", "linux" "macos", "web", "windows".
 class AppcastConfiguration {
   final List<String>? supportedOS;
   final String? url;
@@ -55,16 +55,32 @@ class AppcastConfiguration {
 Upgrader _sharedInstance = Upgrader();
 
 /// A class to configure the upgrade dialog.
-class Upgrader {
+class Upgrader with WidgetsBindingObserver {
+  Upgrader({
+    this.appcastConfig,
+    this.appcast,
+    this.messages,
+    this.debugDisplayAlways = false,
+    this.debugDisplayOnce = false,
+    this.debugLogging = false,
+    this.durationUntilAlertAgain = const Duration(days: 3),
+    this.willDisplayUpgrade,
+    http.Client? client,
+    this.countryCode,
+    this.languageCode,
+    this.minAppVersion,
+    UpgraderOS? upgraderOS,
+  })  : client = client ?? http.Client(),
+        upgraderOS = upgraderOS ?? UpgraderOS() {
+    if (debugLogging) print("upgrader: instantiated.");
+  }
+
   /// Provide an Appcast that can be replaced for mock testing.
   final Appcast? appcast;
 
   /// The appcast configuration ([AppcastConfiguration]) used by [Appcast].
   /// When an appcast is configured for iOS, the iTunes lookup is not used.
   final AppcastConfiguration? appcastConfig;
-
-  /// Can alert dialog be dismissed on tap outside of the alert dialog. Not used by [UpgradeCard]. (default: false)
-  bool canDismissDialog;
 
   /// Provide an HTTP Client that can be replaced for mock testing.
   final http.Client client;
@@ -84,54 +100,18 @@ class Upgrader {
   /// Enable print statements for debugging.
   bool debugLogging;
 
-  /// The upgrade dialog style. Used only on UpgradeAlert. (default: material)
-  UpgradeDialogStyle dialogStyle;
-
   /// Duration until alerting user again
   final Duration durationUntilAlertAgain;
 
   /// The localized messages used for display in upgrader.
-  UpgraderMessages messages;
+  UpgraderMessages? messages;
 
   /// The minimum app version supported by this app. Earlier versions of this app
   /// will be forced to update to the current version. Optional.
   String? minAppVersion;
 
-  /// Called when the ignore button is tapped or otherwise activated.
-  /// Return false when the default behavior should not execute.
-  BoolCallback? onIgnore;
-
-  /// Called when the later button is tapped or otherwise activated.
-  /// Return false when the default behavior should not execute.
-  BoolCallback? onLater;
-
-  /// Called when the update button is tapped or otherwise activated.
-  /// Return false when the default behavior should not execute.
-  BoolCallback? onUpdate;
-
-  /// The [TargetPlatform] that identifies the platform on which the package is
-  /// currently executing. Defaults to [defaultTargetPlatform]. Note that
-  /// [TargetPlatform] does not include web, but includes mobile and desktop.
-  /// This parameter is normally used to change the target platform during testing.
-  final TargetPlatform platform;
-
-  /// Called when the user taps outside of the dialog and [canDismissDialog]
-  /// is false. Also called when the back button is pressed. Return true for
-  /// the screen to be popped. Not used by [UpgradeCard].
-  BoolCallback? shouldPopScope;
-
-  /// Hide or show Ignore button on dialog (default: true)
-  bool showIgnore;
-
-  /// Hide or show Later button on dialog (default: true)
-  bool showLater;
-
-  /// Hide or show release notes (default: true)
-  bool showReleaseNotes;
-
-  /// The text style for the cupertino dialog buttons. Used only for
-  /// [UpgradeDialogStyle.cupertino]. Optional.
-  TextStyle? cupertinoButtonTextStyle;
+  /// Provides information on which OS this code is running on.
+  final UpgraderOS upgraderOS;
 
   /// Called when [Upgrader] determines that an upgrade may or may not be
   /// displayed. The [value] parameter will be true when it should be displayed,
@@ -139,10 +119,6 @@ class Upgrader {
   /// is logging metrics for your app.
   WillDisplayUpgradeCallback? willDisplayUpgrade;
 
-  /// The target operating system.
-  final String _operatingSystem = UpgradeIO.operatingSystem;
-
-  bool _displayed = false;
   bool _initCalled = false;
   PackageInfo? _packageInfo;
 
@@ -160,54 +136,37 @@ class Upgrader {
   /// Track the initialization future so that [initialize] can be called multiple times.
   Future<bool>? _futureInit;
 
-  final notInitializedExceptionMessage =
-      'initialize() not called. Must be called first.';
+  /// A stream that provides a new value each time an evaluation should be performed.
+  /// The values will always be null or true.
+  Stream<UpgraderEvaluateNeed> get evaluationStream => _streamController.stream;
+  final _streamController = StreamController<UpgraderEvaluateNeed>.broadcast();
 
-  Upgrader({
-    this.appcastConfig,
-    this.appcast,
-    UpgraderMessages? messages,
-    this.debugDisplayAlways = false,
-    this.debugDisplayOnce = false,
-    this.debugLogging = false,
-    this.durationUntilAlertAgain = const Duration(days: 3),
-    this.onIgnore,
-    this.onLater,
-    this.onUpdate,
-    this.shouldPopScope,
-    this.willDisplayUpgrade,
-    http.Client? client,
-    this.showIgnore = true,
-    this.showLater = true,
-    this.showReleaseNotes = true,
-    this.canDismissDialog = false,
-    this.countryCode,
-    this.languageCode,
-    this.minAppVersion,
-    this.dialogStyle = UpgradeDialogStyle.material,
-    this.cupertinoButtonTextStyle,
-    TargetPlatform? platform,
-  })  : client = client ?? http.Client(),
-        messages = messages ?? UpgraderMessages(),
-        platform = platform ?? defaultTargetPlatform {
-    if (debugLogging) print("upgrader: instantiated.");
-  }
+  /// An evaluation should be performed.
+  bool get evaluationReady => _evaluationReady;
+  bool _evaluationReady = false;
 
   /// A shared instance of [Upgrader].
   static Upgrader get sharedInstance => _sharedInstance;
+
+  static const notInitializedExceptionMessage =
+      'upgrader: initialize() not called. Must be called first.';
+
+  String? get currentAppStoreListingURL => _appStoreListingURL;
+
+  String? get currentAppStoreVersion => _appStoreVersion;
+
+  String? get currentInstalledVersion => _installedVersion;
+
+  String? get releaseNotes => _releaseNotes;
 
   void installPackageInfo({PackageInfo? packageInfo}) {
     _packageInfo = packageInfo;
     _initCalled = false;
   }
 
-  void installAppStoreVersion(String version) {
-    _appStoreVersion = version;
-  }
+  void installAppStoreVersion(String version) => _appStoreVersion = version;
 
-  void installAppStoreListingURL(String url) {
-    _appStoreListingURL = url;
-  }
+  void installAppStoreListingURL(String url) => _appStoreListingURL = url;
 
   /// Initialize [Upgrader] by getting saved preferences, getting platform package info, and getting
   /// released version info.
@@ -227,28 +186,20 @@ class Upgrader {
       }
       _initCalled = true;
 
-      if (messages.languageCode.isEmpty) {
-        print('upgrader: error -> languageCode is empty');
-      } else if (debugLogging) {
-        print('upgrader: languageCode: ${messages.languageCode}');
-      }
-
-      await _getSavedPrefs();
+      await getSavedPrefs();
 
       if (debugLogging) {
         print('upgrader: default operatingSystem: '
-            '${UpgradeIO.operatingSystem} ${UpgradeIO.operatingSystemVersion}');
-        print('upgrader: operatingSystem: $_operatingSystem');
-        print('upgrader: platform: $platform');
-        print('upgrader: defaultTargetPlatform: $defaultTargetPlatform');
+            '${upgraderOS.operatingSystem} ${upgraderOS.operatingSystemVersion}');
+        print('upgrader: operatingSystem: ${upgraderOS.operatingSystem}');
         print('upgrader: '
-            'isAndroid: ${UpgradeIO.isAndroid}, '
-            'isIOS: ${UpgradeIO.isIOS}, '
-            'isLinux: ${UpgradeIO.isLinux}, '
-            'isMacOS: ${UpgradeIO.isMacOS}, '
-            'isWindows: ${UpgradeIO.isWindows}, '
-            'isFuchsia: ${UpgradeIO.isFuchsia}, '
-            'isWeb: ${UpgradeIO.isWeb}');
+            'isAndroid: ${upgraderOS.isAndroid}, '
+            'isIOS: ${upgraderOS.isIOS}, '
+            'isLinux: ${upgraderOS.isLinux}, '
+            'isMacOS: ${upgraderOS.isMacOS}, '
+            'isWindows: ${upgraderOS.isWindows}, '
+            'isFuchsia: ${upgraderOS.isFuchsia}, '
+            'isWeb: ${upgraderOS.isWeb}');
       }
 
       if (_packageInfo == null) {
@@ -263,78 +214,59 @@ class Upgrader {
 
       _installedVersion = _packageInfo!.version;
 
-      await _updateVersionInfo();
+      await updateVersionInfo();
+
+      // Add an observer of application events.
+      WidgetsBinding.instance.addObserver(this);
+
+      _evaluationReady = true;
+
+      /// Trigger the stream to indicate an evaluation should be performed.
+      /// The value will always be true.
+      _streamController.add(true);
 
       return true;
     });
     return _futureInit!;
   }
 
-  Future<bool> lightWeightInitialize() async {
-    if (_packageInfo == null) {
-      _packageInfo = await PackageInfo.fromPlatform();
-      if (debugLogging) {
-        print(
-            'upgrader: package info packageName: ${_packageInfo!.packageName}');
-        print('upgrader: package info appName: ${_packageInfo!.appName}');
-        print('upgrader: package info version: ${_packageInfo!.version}');
-      }
-    }
-
-    await lightWeightUpdateVersionInfo();
-
-    _installedVersion = _packageInfo!.version;
-    return true;
+  /// Remove any resources allocated.
+  void dispose() {
+    // Remove the observer of application events.
+    WidgetsBinding.instance.removeObserver(this);
   }
 
-  Future<bool> lightWeightUpdateVersionInfo() async {
-    if (_packageInfo == null || _packageInfo!.packageName.isEmpty) {
+  /// Handle application events.
+  @override
+  Future<void> didChangeAppLifecycleState(AppLifecycleState state) async {
+    super.didChangeAppLifecycleState(state);
+
+    // When app has resumed from background.
+    if (state == AppLifecycleState.resumed) {
+      await updateVersionInfo();
+
+      /// Trigger the stream to indicate another evaluation should be performed.
+      /// The value will always be true.
+      _streamController.add(true);
+    }
+  }
+
+  Future<bool> _isHmsDevice() async {
+    GooglePlayServicesAvailability availability = await GoogleApiAvailability
+        .instance
+        .checkGooglePlayServicesAvailability();
+    if (Platform.isIOS) {
+      return false;
+    } else if (Platform.isAndroid) {
+      return availability == GooglePlayServicesAvailability.serviceInvalid;
+    } else {
       return false;
     }
-
-    // The  country code of the locale, defaulting to `US`.
-    final country = countryCode ?? findCountryCode();
-    if (debugLogging) {
-      print('upgrader: countryCode: $country');
-    }
-
-    // Gets Android version either from Google Play Store or Huawei App Gallery
-    // depending on whether it is a HMS device
-    // Gets iOS version from App Store
-
-    if (platform == TargetPlatform.android) {
-      final isHms = await _isHmsDevice();
-
-      if (isHms) {
-        await _getHmsStoreVersion();
-      } else {
-        await _getAndroidStoreVersion(country: country);
-      }
-    } else if (platform == TargetPlatform.iOS) {
-      final iTunes = ITunesSearchAPI();
-      iTunes.client = client;
-      final response = await (iTunes.lookupByBundleId(_packageInfo!.packageName,
-          country: country));
-
-      if (response != null) {
-        _appStoreVersion ??= ITunesResults.version(response);
-        _appStoreListingURL ??= ITunesResults.trackViewUrl(response);
-        _releaseNotes ??= ITunesResults.releaseNotes(response);
-        final mav = ITunesResults.minAppVersion(response);
-        if (mav != null) {
-          minAppVersion = mav.toString();
-          if (debugLogging) {
-            print('upgrader: ITunesResults.minAppVersion: $minAppVersion');
-          }
-        }
-      }
-    }
-    return true;
   }
 
-  Future<bool> _updateVersionInfo() async {
+  Future<bool> updateVersionInfo() async {
     // If there is an appcast for this platform
-    if (_isAppcastThisPlatform()) {
+    if (isAppcastThisPlatform()) {
       if (debugLogging) {
         print('upgrader: appcast is available for this platform');
       }
@@ -370,8 +302,8 @@ class Upgrader {
           _isCriticalUpdate = false;
         }
 
-        _appStoreVersion ??= bestItem.versionString;
-        _appStoreListingURL ??= bestItem.fileURL;
+        _appStoreVersion = bestItem.versionString;
+        _appStoreListingURL = bestItem.fileURL;
         _releaseNotes = bestItem.itemDescription;
       }
     } else {
@@ -393,20 +325,25 @@ class Upgrader {
 
       // Get Android version from Google Play Store, or
       // get iOS version from iTunes Store.
-      if (platform == TargetPlatform.android) {
-        await _getAndroidStoreVersion(country: country, language: language);
-      } else if (platform == TargetPlatform.iOS) {
+      if (upgraderOS.isAndroid) {
+        final isHms = await _isHmsDevice();
+        if (isHms) {
+          await _getHmsStoreVersion();
+        } else {
+          await getAndroidStoreVersion(country: country, language: language);
+        }
+      } else if (upgraderOS.isIOS) {
         final iTunes = ITunesSearchAPI();
-        iTunes.debugEnabled = debugLogging;
+        iTunes.debugLogging = debugLogging;
         iTunes.client = client;
         final response = await (iTunes
             .lookupByBundleId(_packageInfo!.packageName, country: country));
 
         if (response != null) {
-          _appStoreVersion ??= ITunesResults.version(response);
-          _appStoreListingURL ??= ITunesResults.trackViewUrl(response);
-          _releaseNotes ??= ITunesResults.releaseNotes(response);
-          final mav = ITunesResults.minAppVersion(response);
+          _appStoreVersion = iTunes.version(response);
+          _appStoreListingURL = iTunes.trackViewUrl(response);
+          _releaseNotes ??= iTunes.releaseNotes(response);
+          final mav = iTunes.minAppVersion(response);
           if (mav != null) {
             minAppVersion = mav.toString();
             if (debugLogging) {
@@ -435,19 +372,19 @@ class Upgrader {
   }
 
   /// Android info is fetched by parsing the html of the app store page.
-  Future<bool?> _getAndroidStoreVersion(
+  Future<bool?> getAndroidStoreVersion(
       {String? country, String? language}) async {
     final id = _packageInfo!.packageName;
     final playStore = PlayStoreSearchAPI(client: client);
-    playStore.debugEnabled = debugLogging;
+    playStore.debugLogging = debugLogging;
     final response =
         await (playStore.lookupById(id, country: country, language: language));
     if (response != null) {
-      _appStoreVersion ??= PlayStoreResults.version(response);
+      _appStoreVersion ??= playStore.version(response);
       _appStoreListingURL ??=
           playStore.lookupURLById(id, language: language, country: country);
-      _releaseNotes ??= PlayStoreResults.releaseNotes(response);
-      final mav = PlayStoreResults.minAppVersion(response);
+      _releaseNotes ??= playStore.releaseNotes(response);
+      final mav = playStore.minAppVersion(response);
       if (mav != null) {
         minAppVersion = mav.toString();
         if (debugLogging) {
@@ -459,20 +396,7 @@ class Upgrader {
     return true;
   }
 
-  Future<bool> _isHmsDevice() async {
-    GooglePlayServicesAvailability availability = await GoogleApiAvailability
-        .instance
-        .checkGooglePlayServicesAvailability();
-    if (platform == TargetPlatform.iOS) {
-      return false;
-    } else if (platform == TargetPlatform.android) {
-      return availability == GooglePlayServicesAvailability.serviceInvalid;
-    } else {
-      return false;
-    }
-  }
-
-  bool _isAppcastThisPlatform() {
+  bool isAppcastThisPlatform() {
     if (appcastConfig == null ||
         appcastConfig!.url == null ||
         appcastConfig!.url!.isEmpty) {
@@ -484,66 +408,65 @@ class Upgrader {
     // When there are no supported OSes listed, they are all supported.
     var supported = true;
     if (appcastConfig!.supportedOS != null) {
-      supported = appcastConfig!.supportedOS!.contains(_operatingSystem);
+      supported =
+          appcastConfig!.supportedOS!.contains(upgraderOS.operatingSystem);
     }
     return supported;
   }
 
-  bool _verifyInit() {
+  bool verifyInit() {
     if (!_initCalled) {
-      throw (notInitializedExceptionMessage);
+      throw ('upgrader: initialize() not called. Must be called first.');
     }
     return true;
   }
 
   String appName() {
-    _verifyInit();
+    verifyInit();
     return _packageInfo?.appName ?? '';
   }
 
-  String? currentAppStoreListingURL() => _appStoreListingURL;
-
-  String? currentAppStoreVersion() => _appStoreVersion;
-
-  String? currentInstalledVersion() => _installedVersion;
-
-  String? get releaseNotes => _releaseNotes;
-
-  String message() {
+  String body(UpgraderMessages messages) {
     var msg = messages.message(UpgraderMessage.body)!;
     msg = msg.replaceAll('{{appName}}', appName());
     msg = msg.replaceAll(
-        '{{currentAppStoreVersion}}', currentAppStoreVersion() ?? '');
+        '{{currentAppStoreVersion}}', currentAppStoreVersion ?? '');
     msg = msg.replaceAll(
-        '{{currentInstalledVersion}}', currentInstalledVersion() ?? '');
+        '{{currentInstalledVersion}}', currentInstalledVersion ?? '');
     return msg;
   }
 
-  /// Only called by [UpgradeAlert].
-  void checkVersion({
-    required BuildContext context,
-    required bool isDark,
-  }) {
-    if (!_displayed) {
-      final shouldDisplay = shouldDisplayUpgrade();
-      if (debugLogging) {
-        print(
-            'upgrader: shouldDisplayReleaseNotes: ${shouldDisplayReleaseNotes()}');
+  /// Determine which [UpgraderMessages] object to use. It will be either the one passed
+  /// to [Upgrader], or one based on the app locale.
+  UpgraderMessages determineMessages(BuildContext context) {
+    {
+      late UpgraderMessages appMessages;
+      if (messages != null) {
+        appMessages = messages!;
+      } else {
+        String? languageCode;
+        try {
+          // Get the current locale in the app.
+          final locale = Localizations.localeOf(context);
+          // Get the current language code in the app.
+          languageCode = locale.languageCode;
+          if (debugLogging) {
+            print('upgrader: current locale: $locale');
+          }
+        } catch (e) {
+          // ignored, really.
+        }
+
+        appMessages = UpgraderMessages(code: languageCode);
       }
-      if (shouldDisplay) {
-        _displayed = true;
-        Future.delayed(
-          const Duration(milliseconds: 0),
-          () {
-            _showDialog(
-              context: context,
-              isDark: isDark,
-              releaseNotes: shouldDisplayReleaseNotes() ? _releaseNotes : null,
-              canDismissDialog: canDismissDialog,
-            );
-          },
-        );
+
+      if (appMessages.languageCode.isEmpty) {
+        print('upgrader: error -> languageCode is empty');
+      } else if (debugLogging) {
+        print('upgrader: languageCode: ${appMessages.languageCode}');
       }
+
+      return appMessages;
     }
   }
 
@@ -561,12 +484,6 @@ class Upgrader {
       print('upgrader: hasAlerted: $_hasAlerted');
     }
 
-    // If installed version is below minimum app version, or is a critical update,
-    // disable ignore and later buttons.
-    if (isBlocked) {
-      showIgnore = false;
-      showLater = false;
-    }
     bool rv = true;
     if (debugDisplayAlways || (debugDisplayOnce && !_hasAlerted)) {
       rv = true;
@@ -602,7 +519,9 @@ class Upgrader {
         final installedVersion = Version.parse(_installedVersion!);
         rv = installedVersion < minVersion;
       } catch (e) {
-        print(e);
+        if (debugLogging) {
+          print(e);
+        }
       }
     }
     return rv;
@@ -641,24 +560,20 @@ class Upgrader {
       return false;
     }
 
-    if (_updateAvailable == null) {
-      try {
-        final appStoreVersion = Version.parse(_appStoreVersion!);
-        final installedVersion = Version.parse(_installedVersion!);
+    try {
+      final appStoreVersion = Version.parse(_appStoreVersion!);
+      final installedVersion = Version.parse(_installedVersion!);
 
-        final available = appStoreVersion > installedVersion;
-        _updateAvailable = available ? _appStoreVersion : null;
-      } on Exception catch (e) {
+      final available = appStoreVersion > installedVersion;
+      _updateAvailable = available ? _appStoreVersion : null;
+    } on Exception catch (e) {
+      if (debugLogging) {
         print('upgrader: isUpdateAvailable: $e');
       }
     }
     final isAvailable = _updateAvailable != null;
     if (debugLogging) print('upgrader: isUpdateAvailable: $isAvailable');
     return isAvailable;
-  }
-
-  bool shouldDisplayReleaseNotes() {
-    return showReleaseNotes && (_releaseNotes?.isNotEmpty ?? false);
   }
 
   /// Determine the current country code, either from the context, or
@@ -693,116 +608,6 @@ class Upgrader {
     return code;
   }
 
-  void _showDialog({
-    required BuildContext context,
-    required String? releaseNotes,
-    required bool canDismissDialog,
-    required bool isDark,
-  }) {
-    if (debugLogging) {
-      print('upgrader: showDialog message: $message');
-      print('upgrader: showDialog releaseNotes: $releaseNotes');
-    }
-
-    // Save the date/time as the last time alerted.
-    saveLastAlerted();
-
-    showDialog(
-      barrierDismissible: canDismissDialog,
-      context: context,
-      builder: (BuildContext context) {
-        return WillPopScope(
-          onWillPop: () async => _shouldPopScope(),
-          child: AppUpdateDialog(
-            isDark: isDark,
-            messages: messages,
-            releaseNotes: releaseNotes,
-            ignoreCallback:
-                showIgnore ? () => onUserIgnored(context, true) : null,
-            laterCallback: showLater ? () => onUserLater(context, true) : null,
-            updateCallback: () => onUserUpdated(context, !blocked()),
-          ),
-        );
-      },
-    );
-  }
-
-  /// Called when the user taps outside of the dialog and [canDismissDialog]
-  /// is false. Also called when the back button is pressed. Return true for
-  /// the screen to be popped. Defaults to false.
-  bool _shouldPopScope() {
-    if (debugLogging) {
-      print('upgrader: onWillPop called');
-    }
-    if (shouldPopScope != null) {
-      final should = shouldPopScope!();
-      if (debugLogging) {
-        print('upgrader: shouldPopScope=$should');
-      }
-      return should;
-    }
-
-    return false;
-  }
-
-  void onUserIgnored(BuildContext context, bool shouldPop) {
-    if (debugLogging) {
-      print('upgrader: button tapped: ignore');
-    }
-
-    // If this callback has been provided, call it.
-    var doProcess = true;
-    if (onIgnore != null) {
-      doProcess = onIgnore!();
-    }
-
-    if (doProcess) {
-      _saveIgnored();
-    }
-
-    if (shouldPop) {
-      popNavigator(context);
-    }
-  }
-
-  void onUserLater(BuildContext context, bool shouldPop) {
-    if (debugLogging) {
-      print('upgrader: button tapped: later');
-    }
-
-    // If this callback has been provided, call it.
-    var doProcess = true;
-    if (onLater != null) {
-      doProcess = onLater!();
-    }
-
-    if (doProcess) {}
-
-    if (shouldPop) {
-      popNavigator(context);
-    }
-  }
-
-  void onUserUpdated(BuildContext context, bool shouldPop) {
-    if (debugLogging) {
-      print('upgrader: button tapped: update now');
-    }
-
-    // If this callback has been provided, call it.
-    var doProcess = true;
-    if (onUpdate != null) {
-      doProcess = onUpdate!();
-    }
-
-    if (doProcess) {
-      _sendUserToAppStore();
-    }
-
-    if (shouldPop) {
-      popNavigator(context);
-    }
-  }
-
   static Future<void> clearSavedSettings() async {
     var prefs = await SharedPreferences.getInstance();
     await prefs.remove('userIgnoredVersion');
@@ -812,12 +617,7 @@ class Upgrader {
     return;
   }
 
-  void popNavigator(BuildContext context) {
-    Navigator.of(context).pop();
-    _displayed = false;
-  }
-
-  Future<bool> _saveIgnored() async {
+  Future<bool> saveIgnored() async {
     var prefs = await SharedPreferences.getInstance();
 
     _userIgnoredVersion = _appStoreVersion;
@@ -837,7 +637,7 @@ class Upgrader {
     return true;
   }
 
-  Future<bool> _getSavedPrefs() async {
+  Future<bool> getSavedPrefs() async {
     var prefs = await SharedPreferences.getInstance();
     final lastTimeAlerted = prefs.getString('lastTimeAlerted');
     if (lastTimeAlerted != null) {
@@ -851,7 +651,7 @@ class Upgrader {
     return true;
   }
 
-  void _sendUserToAppStore() async {
+  void sendUserToAppStore() async {
     if (_appStoreListingURL == null || _appStoreListingURL!.isEmpty) {
       if (debugLogging) {
         print('upgrader: empty _appStoreListingURL');
@@ -866,8 +666,8 @@ class Upgrader {
     if (await canLaunchUrl(Uri.parse(_appStoreListingURL!))) {
       try {
         await launchUrl(Uri.parse(_appStoreListingURL!),
-            mode: UpgradeIO.isAndroid
-                ? LaunchMode.externalApplication
+            mode: upgraderOS.isAndroid
+                ? LaunchMode.externalNonBrowserApplication
                 : LaunchMode.platformDefault);
       } catch (e) {
         if (debugLogging) {
