@@ -1,6 +1,4 @@
-/*
- * Copyright (c) 2019-2023 Larry Aasen. All rights reserved.
- */
+// Copyright (c) 2023 Larry Aasen. All rights reserved.
 
 import 'package:flutter/material.dart';
 import 'package:upgrader/upgrader.dart';
@@ -25,9 +23,9 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Upgrader Card Example',
+      title: 'Upgrader Example',
       home: Scaffold(
-        appBar: AppBar(title: Text('Upgrader Card Example')),
+        appBar: AppBar(title: Text('Upgrader Custom Card Example')),
         body: Container(
           margin: EdgeInsets.only(left: 12.0, right: 12.0),
           child: SingleChildScrollView(
@@ -35,7 +33,7 @@ class MyApp extends StatelessWidget {
               children: [
                 _simpleCard,
                 _simpleCard,
-                UpgradeCard(),
+                MyUpgradeCard(),
                 _simpleCard,
                 _simpleCard,
               ],
@@ -53,4 +51,38 @@ class MyApp extends StatelessWidget {
           child: Center(child: Text('Card')),
         ),
       );
+}
+
+class MyUpgradeCard extends UpgradeCard {
+  MyUpgradeCard({super.upgrader});
+
+  /// Override the [createState] method to provide a custom class
+  /// with overridden methods.
+  @override
+  UpgradeCardState createState() => MyUpgradeCardState();
+}
+
+class MyUpgradeCardState extends UpgradeCardState {
+  @override
+  Widget buildUpgradeCard(BuildContext context, Key? key) {
+    final appMessages = widget.upgrader.determineMessages(context);
+    final title = appMessages.message(UpgraderMessage.title);
+    return Card(
+      color: Colors.greenAccent,
+      child: AlertStyleWidget(
+        actions: [
+          TextButton(
+            child: Text(
+                appMessages.message(UpgraderMessage.buttonTitleUpdate) ?? ''),
+            onPressed: () {
+              widget.upgrader.saveLastAlerted();
+              onUserUpdated();
+            },
+          ),
+        ],
+        content: Text(''),
+        title: Text(title ?? ''),
+      ),
+    );
+  }
 }
