@@ -20,7 +20,9 @@ void main() async {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  MyApp({Key? key}) : super(key: key);
+
+  final upgrader = MyUpgrader();
 
   @override
   Widget build(BuildContext context) {
@@ -29,7 +31,7 @@ class MyApp extends StatelessWidget {
       home: Scaffold(
           appBar: AppBar(title: Text('Upgrader Subclass Example')),
           body: UpgradeAlert(
-            upgrader: MyUpgrader(),
+            upgrader: upgrader,
             child: Center(child: Text('Checking...')),
           )),
     );
@@ -39,11 +41,4 @@ class MyApp extends StatelessWidget {
 /// This class extends / subclasses Upgrader.
 class MyUpgrader extends Upgrader {
   MyUpgrader() : super(debugLogging: true);
-
-  /// This method overrides super class method.
-  @override
-  void popNavigator(BuildContext context) {
-    print('this method overrides popNavigator');
-    super.popNavigator(context);
-  }
 }
