@@ -1,18 +1,17 @@
 // ignore_for_file: constant_identifier_names
 
 /*
- * Copyright (c) 2018-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2023 Larry Aasen. All rights reserved.
  */
 
 import 'dart:convert' show utf8;
 
-import 'package:device_info_plus/device_info_plus.dart';
-import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
 import 'package:version/version.dart';
 import 'package:xml/xml.dart';
 
-import 'upgrade_io.dart';
+import 'upgrade_os.dart';
+import 'upgrade_device.dart';
 
 /// The [Appcast] class is used to download an Appcast, based on the Sparkle
 /// framework by Andy Matuschak.
@@ -20,20 +19,26 @@ import 'upgrade_io.dart';
 /// An Appcast is an RSS feed with one channel that has a collection of items
 /// that each describe one app version.
 class Appcast {
-  /// Provide an HTTP Client that can be replaced for mock testing.
-  http.Client? client;
+  /// Provide an HTTP Client that can be replaced during testing.
+  final http.Client client;
+
+  /// Provide [UpgraderOS] that can be replaced during testing.
+  final UpgraderOS upgraderOS;
+
+  /// Provide [UpgraderDevice] that ca be replaced during testing.
+  final UpgraderDevice upgraderDevice;
 
   Appcast({
-    this.client,
-  }) {
-    client ??= http.Client();
-  }
+    http.Client? client,
+    UpgraderOS? upgraderOS,
+    UpgraderDevice? upgraderDevice,
+  })  : client = client ?? http.Client(),
+        upgraderOS = upgraderOS ?? UpgraderOS(),
+        upgraderDevice = upgraderDevice ?? UpgraderDevice();
 
   /// The items in the Appcast.
   List<AppcastItem>? items;
 
-  late AndroidDeviceInfo _androidInfo;
-  late IosDeviceInfo _iosInfo;
   String? osVersionString;
 
   /// Returns the latest critical item in the Appcast.
@@ -44,7 +49,9 @@ class Appcast {
 
     AppcastItem? bestItem;
     items!.forEach((AppcastItem item) {
-      if (item.hostSupportsItem(osVersion: osVersionString) &&
+      if (item.hostSupportsItem(
+              osVersion: osVersionString,
+              currentPlatform: upgraderOS.current) &&
           item.isCriticalUpdate) {
         if (bestItem == null) {
           bestItem = item;
@@ -73,7 +80,8 @@ class Appcast {
 
     AppcastItem? bestItem;
     items!.forEach((AppcastItem item) {
-      if (item.hostSupportsItem(osVersion: osVersionString)) {
+      if (item.hostSupportsItem(
+          osVersion: osVersionString, currentPlatform: upgraderOS.current)) {
         if (bestItem == null) {
           bestItem = item;
         } else {
@@ -96,7 +104,7 @@ class Appcast {
   Future<List<AppcastItem>?> parseAppcastItemsFromUri(String appCastURL) async {
     http.Response response;
     try {
-      response = await client!.get(Uri.parse(appCastURL));
+      response = await client.get(Uri.parse(appCastURL));
     } catch (e) {
       print('upgrader: parseAppcastItemsFromUri exception: $e');
       return null;
@@ -107,7 +115,7 @@ class Appcast {
 
   /// Parse the Appcast from XML string.
   Future<List<AppcastItem>?> parseAppcastItems(String contents) async {
-    await _getDeviceInfo();
+    osVersionString = await upgraderDevice.getOsVersionString(upgraderOS);
     return parseItemsFromXMLString(contents);
   }
 
@@ -216,40 +224,6 @@ class Appcast {
 
     return items;
   }
-
-  Future<bool> _getDeviceInfo() async {
-    final deviceInfo = DeviceInfoPlugin();
-    if (UpgradeIO.isAndroid) {
-      _androidInfo = await deviceInfo.androidInfo;
-      osVersionString = _androidInfo.version.baseOS;
-    } else if (UpgradeIO.isIOS) {
-      _iosInfo = await deviceInfo.iosInfo;
-      osVersionString = _iosInfo.systemVersion;
-    } else if (UpgradeIO.isMacOS) {
-      final info = await deviceInfo.macOsInfo;
-      final release = info.osRelease;
-
-      // For macOS the release string looks like: Version 13.2.1 (Build 22D68)
-      // We need to parse out the actual OS version number.
-
-      String regExpSource = r"[\w]*[\s]*(?<version>[^\s]+)";
-      final regExp = RegExp(regExpSource, caseSensitive: false);
-      final match = regExp.firstMatch(release);
-      final version = match?.namedGroup('version');
-      osVersionString = version;
-    } else if (UpgradeIO.isWeb) {
-      osVersionString = '0.0.0';
-    }
-
-    // If the OS version string is not valid, don't use it.
-    try {
-      Version.parse(osVersionString!);
-    } catch (e) {
-      osVersionString = null;
-    }
-
-    return true;
-  }
 }
 
 class AppcastItem {
@@ -289,13 +263,13 @@ class AppcastItem {
       ? false
       : tags!.contains(AppcastConstants.ElementCriticalUpdate);
 
-  bool hostSupportsItem({String? osVersion, String? currentPlatform}) {
-    var supported = true;
+  /// Does the host support this item? If so is [osVersion] supported?
+  bool hostSupportsItem({String? osVersion, required String currentPlatform}) {
+    assert(currentPlatform.isNotEmpty);
+    bool supported = true;
     if (osString != null && osString!.isNotEmpty) {
       final platformEnum = 'TargetPlatform.${osString!}';
-      currentPlatform = currentPlatform == null
-          ? defaultTargetPlatform.toString()
-          : 'TargetPlatform.$currentPlatform';
+      currentPlatform = 'TargetPlatform.$currentPlatform';
       supported = platformEnum.toLowerCase() == currentPlatform.toLowerCase();
     }
 
