@@ -5,8 +5,6 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
@@ -15,24 +13,12 @@ import 'package:upgrader/upgrader.dart';
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  setMockDeviceInfo();
-
-  setUp(() async {});
-
-  tearDown(() async {});
-
-  test('testing defaultTargetPlatform', () async {
-    // Platform.operatingSystem can be "macos" or "linux" in a unit test.
-    // defaultTargetPlatform is TargetPlatform.android in a unit test.
-
-    // Flutter testings assumes the platform is android, so verify it.
-    expect(defaultTargetPlatform, equals(TargetPlatform.android));
-  });
-
   /// These tests inspired by:
   ///   https://github.com/sparkle-project/Sparkle/blob/master/Tests/SUAppcastTest.swift
-  test('testing Appcast', () async {
-    final appcast = Appcast();
+  test('testing Appcast defaults', () async {
+    final appcast = Appcast(
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
     expect(appcast.bestItem(), isNull);
     expect(appcast.osVersionString, isNull);
     expect(appcast.items, isNull);
@@ -42,24 +28,30 @@ void main() {
   });
 
   test('testing Appcast file', () async {
-    final appcast = TestAppcast();
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
     var testFile = await getTestFile();
     final items = await appcast.parseAppcastItemsFromFile(testFile);
     validateItems(items!, appcast);
   });
 
-  test('testing Appcast ', () async {
-    final client = await setupMockClient();
-    final appcast = Appcast(client: client);
+  test('testing Appcast client', () async {
+    final client = setupMockClient();
+    final appcast = Appcast(
+        client: client,
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
     final items = await appcast.parseAppcastItemsFromUri(
         'https://sparkle-project.org/test/testappcast.xml');
     validateItems(items!, appcast);
   }, skip: false);
 
-  test(
-      'testing Appcast will prioritize critical version even with lover version. ',
+  test('Appcast will prioritize critical version even with lower version',
       () async {
-    final appcast = TestAppcast();
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
 
     final testFile =
         await getTestFile(filePath: 'test/testappcast_critical.xml');
@@ -80,19 +72,123 @@ void main() {
   test('testing Appcast host', () async {
     final item = AppcastItem();
 
-    expect(item.hostSupportsItem(osVersion: null), equals(true));
-    expect(item.hostSupportsItem(osVersion: ''), equals(true));
-    expect(item.hostSupportsItem(osVersion: '0'), equals(true));
+    expect(item.hostSupportsItem(osVersion: null, currentPlatform: 'android'),
+        equals(true));
+    expect(item.hostSupportsItem(osVersion: '', currentPlatform: 'android'),
+        equals(true));
+    expect(item.hostSupportsItem(osVersion: '0', currentPlatform: 'android'),
+        equals(true));
 
     expect(
         item.hostSupportsItem(
             osVersion:
-                'samsung/hero2ltexx/hero2lte:7.0/NRD90M/G935FXXU2DRB6:user/release-keys'),
+                'samsung/hero2ltexx/hero2lte:7.0/NRD90M/G935FXXU2DRB6:user/release-keys',
+            currentPlatform: 'android'),
         equals(false));
 
-    expect(item.hostSupportsItem(osVersion: '0.1'), equals(true));
+    expect(item.hostSupportsItem(osVersion: '0.1', currentPlatform: 'android'),
+        equals(true));
 
-    expect(item.hostSupportsItem(osVersion: '0.0.1'), equals(true));
+    expect(
+        item.hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+        equals(true));
+  });
+
+  test('Appcast multi Android', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.1.2'));
+  });
+
+  test('Appcast multi Fuchsia', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(fuchsia: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.2.2'));
+  });
+
+  test('Appcast multi iOS', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(ios: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.3.2'));
+  });
+
+  test('Appcast multi Linux', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(linux: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.4.2'));
+  });
+
+  test('Appcast multi macOS', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(macos: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.5.2'));
+  });
+
+  /// The [MockDeviceInfo] does not work properly for windows [WindowsDeviceInfo], so for now this
+  /// test is skipped.
+  test('Appcast multi Windows', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(windows: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.6.2'));
+  }, skip: true);
+
+  test('Appcast multi Web', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(web: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile = await getTestFile(filePath: 'test/testappcastmulti.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem()!;
+    expect(bestItem, isNotNull);
+    expect(bestItem.versionString, equals('2.7.2'));
+  });
+  test('Appcast multi multi enclosure', () async {
+    final appcast = TestAppcast(
+        upgraderOS: MockUpgraderOS(android: true),
+        upgraderDevice: MockUpgraderDevice());
+    var testFile =
+        await getTestFile(filePath: 'test/testappcast-enclosure.xml');
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestItem = appcast.bestItem();
+    expect(bestItem, isNull);
   });
 }
 
@@ -110,7 +206,9 @@ void validateItems(List<AppcastItem> items, Appcast appcast) {
   expect(items[0].maximumSystemVersion, isNull);
   expect(items[0].minimumSystemVersion, isNull);
   expect(items[0].versionString, equals('2.0'));
-  expect(items[0].hostSupportsItem(osVersion: '0.0.1'), equals(true));
+  expect(
+      items[0].hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+      equals(true));
   expect(items[0].osString, isNull);
 
   expect(items[1].title, equals('Version 3.0'));
@@ -122,7 +220,9 @@ void validateItems(List<AppcastItem> items, Appcast appcast) {
   expect(items[1].maximumSystemVersion, isNull);
   expect(items[1].minimumSystemVersion, isNull);
   expect(items[1].versionString, equals('3.0'));
-  expect(items[1].hostSupportsItem(osVersion: '0.0.1'), equals(true));
+  expect(
+      items[1].hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+      equals(true));
   expect(items[1].osString, equals('android'));
 
   expect(items[2].title, equals('Version 4.0'));
@@ -134,7 +234,9 @@ void validateItems(List<AppcastItem> items, Appcast appcast) {
   expect(items[2].maximumSystemVersion, isNull);
   expect(items[2].minimumSystemVersion, equals('17.0.0'));
   expect(items[2].versionString, equals('4.0'));
-  expect(items[2].hostSupportsItem(osVersion: '0.0.1'), equals(false));
+  expect(
+      items[2].hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+      equals(false));
   expect(items[2].hostSupportsItem(osVersion: '17.0.1', currentPlatform: 'iOS'),
       equals(true));
   expect(items[2].osString, equals('iOS'));
@@ -147,14 +249,20 @@ void validateItems(List<AppcastItem> items, Appcast appcast) {
   expect(items[3].maximumSystemVersion, equals('2.0.0'));
   expect(items[3].minimumSystemVersion, isNull);
   expect(items[3].versionString, equals('5.0'));
-  expect(items[3].hostSupportsItem(osVersion: '0.0.1'), equals(true));
-  expect(items[3].hostSupportsItem(osVersion: '2.0.1'), equals(false));
+  expect(
+      items[3].hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+      equals(true));
+  expect(
+      items[3].hostSupportsItem(osVersion: '2.0.1', currentPlatform: 'android'),
+      equals(false));
   expect(items[3].osString, isNull);
 
   final bestItem = appcast.bestItem()!;
   expect(bestItem, isNotNull);
   expect(bestItem.versionString, equals('5.0'));
-  expect(bestItem.hostSupportsItem(osVersion: '0.0.1'), equals(true));
+  expect(
+      bestItem.hostSupportsItem(osVersion: '0.0.1', currentPlatform: 'android'),
+      equals(true));
   expect(bestItem.osString, isNull);
 }
 
@@ -167,8 +275,7 @@ Future<File> getTestFile({String filePath = 'test/testappcast.xml'}) async {
   return testFile;
 }
 
-Future<http.Client> setupMockClient(
-    {String filePath = 'test/testappcast.xml'}) async {
+http.Client setupMockClient({String filePath = 'test/testappcast.xml'}) {
   // Use a mock to return a successful response when it calls the
   // provided http.Client.
 
@@ -186,44 +293,11 @@ Future<http.Client> setupMockClient(
 }
 
 class TestAppcast extends Appcast {
+  TestAppcast({super.client, super.upgraderOS, super.upgraderDevice});
+
   /// Load the Appcast from [file].
   Future<List<AppcastItem>?> parseAppcastItemsFromFile(File file) async {
     final contents = await file.readAsString();
     return parseAppcastItems(contents);
   }
-}
-
-void setMockDeviceInfo() {
-  const channel = MethodChannel('dev.fluttercommunity.plus/device_info');
-
-  handler(MethodCall methodCall) async {
-    print('setMockDeviceInfo.methodCall: ${methodCall.method}');
-
-    switch (methodCall.method) {
-      case 'getMacosDeviceInfo':
-      case 'getDeviceInfo':
-        return <String, dynamic>{
-          'computerName': '',
-          'hostName': '',
-          'arch': '',
-          'model': '',
-          'kernelVersion': '',
-          'osRelease': 'Version 13.2.1 (Build 22D68)',
-          'majorVersion': 13,
-          'minorVersion': 2,
-          'patchVersion': 1,
-          'activeCPUs': 0,
-          'memorySize': 0,
-          'cpuFrequency': 0,
-          'systemGUID': '',
-        };
-
-      default:
-        assert(false);
-        return null;
-    }
-  }
-
-  TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-      .setMockMethodCallHandler(channel, handler);
 }
