@@ -12,7 +12,8 @@ import 'mock_play_store_client.dart';
 /// Helper method
 String? pmav(Document response,
     {String tagRES = r'\[\:mav\:[\s]*(?<version>[^\s]+)[\s]*\]'}) {
-  final mav = PlayStoreResults.minAppVersion(response, tagRegExpSource: tagRES);
+  final mav =
+      PlayStoreSearchAPI().minAppVersion(response, tagRegExpSource: tagRES);
   return mav?.toString();
 }
 
@@ -26,13 +27,17 @@ void main() {
     expect(Version.parse('1.2.3+1').toString(), '1.2.3+1');
     expect(Version.parse('0.0.0').toString(), '0.0.0');
     expect(Version.parse('0.0.0+1').toString(), '0.0.0+1');
+
+    final version1 = Version.parse('1.2.3+1');
+    final version2 = Version.parse('1.2.3+2');
+    expect(version1 == version2, isTrue);
   }, skip: false);
 
   test('testing PlayStoreSearchAPI properties', () async {
     final playStore = PlayStoreSearchAPI();
-    expect(playStore.debugEnabled, equals(false));
-    playStore.debugEnabled = true;
-    expect(playStore.debugEnabled, equals(true));
+    expect(playStore.debugLogging, equals(false));
+    playStore.debugLogging = true;
+    expect(playStore.debugLogging, equals(true));
     expect(playStore.playStorePrefixURL.length, greaterThan(0));
 
     expect(
@@ -50,9 +55,9 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '1.23.0');
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '1.23.0');
 
     expect(await playStore.lookupById('com.not.a.valid.application'), isNull);
 
@@ -74,23 +79,23 @@ void main() {
     expect(() => playStore.lookupURLById(''), throwsAssertionError);
     expect(
         playStore.lookupURLById('com.testing.test1')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=en&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: null)!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: '')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', country: 'IN')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN&hl=en&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=IN&hl=en&_cb=17'),
         equals(true));
     expect(
         playStore.lookupURLById('com.testing.test1', language: 'es')!.startsWith(
-            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=es&_cb=16'),
+            'https://play.google.com/store/apps/details?id=com.testing.test1&gl=US&hl=es&_cb=17'),
         equals(true));
     expect(
         playStore
@@ -109,10 +114,10 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '2.3.0');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '2.3.0');
+    expect(playStore.description(response)?.length, greaterThan(10));
     expect(
         pmav(response,
             tagRES:
@@ -130,10 +135,10 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
-        'Minor updates and improvements.');
-    expect(PlayStoreResults.version(response), '2.0.2');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
+    expect(
+        playStore.releaseNotes(response!), 'Minor updates and improvements.');
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
   }, skip: false);
 
   test('testing release notes <br>', () async {
@@ -144,10 +149,10 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
+    expect(playStore.releaseNotes(response!),
         'Minor updates and improvements.\nAgain.\nAgain.');
-    expect(PlayStoreResults.version(response), '2.0.2');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
   }, skip: false);
 
   test('testing release notes <br> 2', () async {
@@ -158,16 +163,10 @@ void main() {
     expect(response, isNotNull);
     expect(response, isInstanceOf<Document>());
 
-    expect(PlayStoreResults.releaseNotes(response!),
+    expect(playStore.releaseNotes(response!),
         'Minor updates and improvements.\nAgain.\nAgain.');
-    expect(PlayStoreResults.version(response), '2.0.2');
-    expect(PlayStoreResults.description(response)?.length, greaterThan(10));
-  }, skip: false);
-
-  test('testing PlayStoreResults', () async {
-    expect(PlayStoreResults(), isNotNull);
-    expect(PlayStoreResults.releaseNotes(Document()), isNull);
-    expect(PlayStoreResults.version(Document()), isNull);
+    expect(playStore.version(response), '2.0.2');
+    expect(playStore.description(response)?.length, greaterThan(10));
   }, skip: false);
 
   /// Helper method
